Move distillation utils into main script
